requirements.txt
spotipy
streamlit
pandas
python-dotenv